worked on dialogue editor and dialogue manager. The editor more or less fully works now just have to format it and make sure to put some validation in place so that for example an ending node can't be attached as a parent.
Next step is to finish the dialogue manager and make sure that the choices proccess properly which is what I am working on now. The problem was that if the choice was clicked on before the text finished displaying the text area would mess up. I changed it and then the choice buttons dispaeared. So I have to fix that next.